% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_heatmap.R
\name{abundance_heatmap}
\alias{abundance_heatmap}
\title{Create a heatmap of the out_table from a
phyloseq-object.}
\usage{
abundance_heatmap(phyloseq_obj, classification = NULL, 
  treatment =  NULL, subset = NULL, transformation = 'none', 
  colors = 'default', treatment_labels = NULL, sample_labels = NULL,
  classification_labels= NULL)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object.}

\item{classification}{The level of taxonomy to examine. Must be a column name
from the tax_table of the phyloseq_object.
\code{\link[phyloseq:tax_table]{tax_table}}.}

\item{treatment}{Column name as a string, or vector of strings, from the
\code{\link[phyloseq:sample_data]{sample_data}}.}

\item{subset}{A level within the \code{treatment}. Multiple levels can be 
given as a vector.}

\item{transformation}{Transformation to be used on the data. "none",
"relative_abundance", "log", "log10", "log1p", "log2", "asn", "atanh",
"boxcox", "exp", "identity", "logit", "probability", "probit",
"reciprocal", "reverse" and "sqrt"}

\item{colors}{Name of a color set from the
\link[=RColorBrewer]{RColorBrewer} package or a vector palette of R-accepted
colors. The default is an adaption of the palette from 
\url{https://www.nature.com/articles/nmeth.1618}}

\item{treatment_labels}{a vector of names to be used as labels for
treatments/facets, in the order they appear in the figure.}

\item{sample_labels}{a vector of names to be used as labels for Samples, 
in the order they appear in the figure.}

\item{classification_labels}{a vector of names to be used as labels for the
taxonomic classifications, in the order they appear in the figure.}
}
\value{
ggplot-object
}
\description{
Uses a \code{\link[phyloseq]{phyloseq-class}} object as input and
creates a ggplot-heatmap of the abundances across samples.
The default color choice is the viridis palette.
}
\examples{
data(GlobalPatterns)
abundance_heatmap(GlobalPatterns, classification = "Phylum",
 treatment = "SampleType", transformation = "log2")
}
