% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_phyloseq_ggplot.R
\name{abundance_heatmap_ggplot}
\alias{abundance_heatmap_ggplot}
\title{Create a ggplot object of the heatmap of the abundance table from a phyloseq object. Function from the phylosmith-package.}
\usage{
abundance_heatmap_ggplot(phyloseq_obj, classification = 'none', treatment, subset = NULL,
transformation = 'none', colors = 'default')
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object created with the \link[=phyloseq]{phyloseq} package (must contain \code{\link[phyloseq:sample_data]{sample_data()}}).}

\item{classification}{Column name or number in the \code{\link[phyloseq:tax_table]{tax_table}}.}

\item{treatment}{Column name or number, or vector of, in the \code{\link[phyloseq:sample_data]{sample_data}}.}

\item{subset}{If taxa not needed to be seen in all \code{treatment}, then will subset to treatments containing this string.}

\item{transformation}{Transforms the data. "none", "relative_abundance", "log", "log10", "log1p", "log2", "asn", "atanh", "boxcox", "exp", "identity", "logit", "probability", "probit", "reciprocal", "reverse" and "sqrt"}

\item{colors}{Name of a color set from the \link[=RColorBrewer]{RColorBrewer} package or a vector palete of R accepted colors.}
}
\description{
This function takes a \code{\link[phyloseq]{phyloseq-class}} object and creates heatmaps of the abundances across samples.
}
