% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_phyloseq_ggplot.R
\name{abundance_lines_ggplot}
\alias{abundance_lines_ggplot}
\title{Create a ggplot object of the abundance table from a phyloseq object. Function from the phylosmith-package.}
\usage{
abundance_lines_ggplot(phyloseq_obj, classification_level, treatment, subset = NULL,
relative_abundance = FALSE, points = TRUE, colors = "Spectral")
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object created with the \link[=phyloseq]{phyloseq} package (must contain \code{\link[phyloseq:sample_data]{sample_data()}}).}

\item{classification_level}{Column name or number in the \code{\link[phyloseq:tax_table]{tax_table}}.}

\item{treatment}{Column name or number, or vector of, in the \code{\link[phyloseq:sample_data]{sample_data}}.}

\item{subset}{If taxa not needed to be seen in all \code{treatment}, then can check only one particular treatment subset, this works for multiple treatment inputs.}

\item{relative_abundance}{If TRUE, transforms the abundance data into relative abundance by sample.}

\item{points}{if TRUE, will diplay the data-points.}

\item{colors}{Name of a color set from the \link[=RColorBrewer]{RColorBrewer} package.}
}
\description{
This function takes a \code{\link[phyloseq]{phyloseq-class}} object and creates line graphs across samples.
}
