% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{alpha_diversity_graph}
\alias{alpha_diversity_graph}
\title{Create a boxplot of the alpha-diversity. Function from the phylosmith-package.}
\usage{
alpha_diversity_graph(phyloseq_obj, treatment = NULL, subset = NULL,
index = 'shannon', colors = 'default')
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object. It
must contain \code{\link[phyloseq:sample_data]{sample_data()}}) with
information about each sample, and it must contain
\code{\link[phyloseq:tax_table]{tax_table()}}) with information about
each taxa/gene.}

\item{treatment}{Column name as a string or number in the
\code{\link[phyloseq:sample_data]{sample_data}}. This can be a vector of
multiple columns and they will be combined into a new column.}

\item{subset}{A factor within the \code{treatment}. This will remove any
samples that to not contain this factor. This can be a vector of multiple
factors to subset on.}

\item{index}{The diversity index to calculate ('shannon', 'simpson', 'invsimpson')}

\item{colors}{Name of a color set from the
\link[=RColorBrewer]{RColorBrewer} package or a vector palette of R-accepted
colors.}
}
\value{
ggplot-object
}
\description{
Inputs a \code{\link[phyloseq]{phyloseq-class}} object and
creates boxplot of the alpha diversity as a ggplot object.
}
\examples{
alpha_diversity_graph(soil_column, index = 'shannon',
treatment = c('Matrix', 'Treatment'), subset = NULL, colors = 'default')
}
