% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arrange_co_occurrence_table}
\alias{arrange_co_occurrence_table}
\title{Arrange co-occurence table}
\usage{
arrange_co_occurrence_table(co_occurrence_table, taxa_of_interest)
}
\arguments{
\item{co_occurrence_table}{A \code{data.frame} in the format from
\code{\link{co_occurrence}}.}

\item{taxa_of_interest}{A \code{vector} containing names of taxa to be
found in either OTU_1 or OTU_2.}
}
\value{
A \code{data.frame} with treatment, otu_1, otu_2, rho, p values.
}
\description{
Arranges the co-occurence table so that taxa of interest are
on the left.
}
\seealso{
\code{\link{co_occurrence}}
}
\author{
Schuyler D. Smith
}
