% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{assign_rank}
\alias{assign_rank}
\title{Assign rank values to a matrix.}
\usage{
assign_rank(count_matrix)
}
\arguments{
\item{count_matrix}{A \code{data.frame}}
}
\value{
A rank matrix
}
\description{
Arranges the co-occurence table so that taxa of interest are
on the left.
}
\author{
Schuyler D. Smith
}
