% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineternal.R
\name{bin}
\alias{bin}
\title{Bin data}
\usage{
bin(data, nbins, labels = NULL)
}
\arguments{
\item{data}{vector of data to bin}

\item{nbins}{number of bins to produce}

\item{labels}{names for the bins}
}
\value{
bins of values
}
\description{
Assigns bins to a vector of values
}
