% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_occurrence.R
\name{bootstrap_rho}
\alias{bootstrap_rho}
\title{Bootstraps the pair-wise Spearman rank co-occurrence, to determine a significant rho-cutoff. phylosmith}
\usage{
bootstrap_rho(phyloseq_obj, treatment, replicates = 'independent', permutations = 100,
cores = 0)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object created with the \link[=phyloseq]{phyloseq} package.}

\item{treatment}{Column name or number, or vector of, in the \code{\link[phyloseq:sample_data]{sample_data}}.}

\item{replicates}{Column name or number, or vector of, in the \code{\link[phyloseq:sample_data]{sample_data}} that indicates which samples are non-independent of each other.}

\item{permutations}{Number of iterations to compute.}

\item{cores}{Number of CPU cores to use for the pair-wise permutations. Default uses all cores available.}
}
\description{
Bootstraps the pair-wise Spearman rank co-occurrence, to determine a significant rho-cutoff.
}
\examples{
data(mock_phyloseq)
bootstrap_rho(mock_phyloseq, treatment = "day", permutations = 10)
}
\seealso{
\code{\link{co_occurrence}}
}
\keyword{nonparametric}
