% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineternal.R
\name{change_labels}
\alias{change_labels}
\title{Change names of factors in graph_data to change labels in plot objects}
\usage{
change_labels(graph_data, treatment = NULL, treatment_labels = NULL,
sample_labels = NULL, classification = NULL, classification_labels = NULL)
}
\arguments{
\item{graph_data}{A \code{\link[data.table:data.table]{data.table()}}) object
containing the melted phyloseq data.}

\item{treatment}{name of the column defined as the treatment.}

\item{treatment_labels}{a vector of names to be used as labels for
treatments/facets.}

\item{sample_labels}{a vector of names to be used as labels for Samples.}

\item{classification}{name of the column defined as the classification.}

\item{classification_labels}{a vector of names to be used as labels for the
taxonomic classifications.}
}
\value{
data.table
}
\description{
Change names of factors in graph_data to change labels in plot objects
}
