% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineternal.R
\name{check_index_treatment}
\alias{check_index_treatment}
\title{Converts numeric values to column names in sample_data.}
\usage{
check_index_treatment(phyloseq_obj, columns)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object. It
must contain \code{\link[phyloseq:sample_data]{sample_data()}}) with
information about each sample, and it must contain
\code{\link[phyloseq:tax_table]{tax_table()}}) with information about each
taxa/gene.}

\item{columns}{Column name as a \code{string} or \code{numeric} in the
\code{\link[phyloseq:sample_data]{sample_data}}. This can be any number of
multiple columns and they will be combined into a new column.}
}
\value{
string
}
\description{
Converts numeric values to column names in sample_data.
}
