% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_export.R
\name{check_numeric_treatment}
\alias{check_numeric_treatment}
\title{Converts numeric values to column names in sample_data.}
\usage{
check_numeric_treatment(phyloseq_obj, treatments)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object. It must contain \code{\link[phyloseq:sample_data]{sample_data()}}) with information about each sample, and it must contain \code{\link[phyloseq:tax_table]{tax_table()}}) with information about each taxa/gene.}

\item{treatments}{Column name as a \code{string} or \code{numeric} in the \code{\link[phyloseq:sample_data]{sample_data}}. This can be a vector of multiple columns and they will be combined into a new column.}
}
\description{
Converts numeric values to column names in sample_data.
}
