% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineternal.R
\name{check_TaR}
\alias{check_TaR}
\title{phyloseq object so taxa are rows.}
\usage{
check_TaR(phyloseq_obj)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object. It
must contain \code{\link[phyloseq:sample_data]{sample_data()}}) with
information about each sample, and it must contain
\code{\link[phyloseq:tax_table]{tax_table()}}) with information about each
taxa/gene.}
}
\value{
phyloseq_obj
}
\description{
phyloseq object so taxa are rows.
}
