% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_ARG_codes.R
\name{classify_ARG_class}
\alias{classify_ARG_class}
\title{classify_ARG_class}
\usage{
classify_ARG_genes(phyloseq_obj, genes, obo = NULL)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object created with the \link[=phyloseq]{phyloseq} package.}

\item{genes}{Column name or number in the \code{\link[phyloseq:tax_table]{tax_table}} where ARGs are as gene names or ARO accession number.}

\item{obo}{an object processed with \code{\link{processOBO}}.}
}
\description{
Classifies ARGs from the \href{https://card.mcmaster.ca/home}{CARD database} using either accession number or gene name.
}
