% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_occurrence.R
\name{co_occurrence}
\alias{co_occurrence}
\alias{FastCoOccur}
\title{pair-wise Spearman rank co-occurrence, written in efficient c++ code. Function from the phylosmith-package.}
\usage{
co_occurrence(phyloseq_obj, treatment = NULL, p = 0.05, cores = 0)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object. It must contain \code{\link[phyloseq:sample_data]{sample_data()}} with information about each sample, and it must contain \code{\link[phyloseq:tax_table]{tax_table()}}) with information about each taxa/gene.}

\item{treatment}{Column name as a string or number in the \code{\link[phyloseq:sample_data]{sample_data}}. This can be a vector of multiple columns and they will be combined into a new column.}

\item{p}{The p-value cutoff. All returned co-occurrences will have a p-value less than or equal to \code{p}.}

\item{cores}{Number of CPU cores to use for the pair-wise permutations. Default (0) uses max cores available. Parallelization not available for systems running MacOS without openMP configuration.}
}
\description{
A rewrite of the pair-wise Spearman rank co-occurrence routine written by \href{https://github.com/germs-lab/FastCoOccur}{Jin Choi}. The routine has been adapted to integrate with the \code{\link[Rcpp]{Rcpp-package}} API.
}
\seealso{
\code{\link{bootstrap_rho}} \code{\link{phylosmith}}
}
\keyword{nonparametric}
