% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{co_occurrence_Rcpp}
\alias{co_occurrence_Rcpp}
\title{Co-occurrence calculation}
\usage{
co_occurrence_Rcpp(otu_table, treatment_indices,
treatment_names, rho_cutoff, p_cutoff, ncores)
}
\arguments{
\item{otu_table}{An \code{otu_table} in the format from
\code{\link[phyloseq:otu_table]{phyloseq}}}

\item{treatment_indices}{A \code{list} with c++ indices for the
\code{treatment_names} corresponding to which treatment each column in the
\code{otu_table} belongs to.}

\item{treatment_names}{A \code{Vector} containing the treatment names
corresponding to the \code{treatment_indices}.}

\item{rho_cutoff}{\code{double} representing the minimum \code{rho-value}
accepted for the correlation to be returned.}

\item{p_cutoff}{\code{double} representing the maximum \code{p-value}
accepted for the correlation to be returned.}

\item{ncores}{An \code{int} for how many cores to use to multithread the
calculations.}
}
\value{
A \code{data.frame} with treatment, otu_1, otu_2, rho, p values.
}
\description{
Calculate the pair-wise Spearman rank correlation.
}
\seealso{
\code{\link{co_occurrence}}
}
\author{
Schuyler D. Smith
}
