% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_occurrence.R
\name{co_occurrence_rho}
\alias{co_occurrence_rho}
\title{pair-wise Spearman rank co-occurrence, written in efficient c++ code.}
\usage{
co_occurrence_rho(phyloseq_obj, treatment, cores = 0)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object created with the \link[=phyloseq]{phyloseq} package.}

\item{treatment}{Column name or number, or vector of, in the \code{\link[phyloseq:sample_data]{sample_data}}.}

\item{cores}{Number of CPU cores to use for the pair-wise permutations. Default uses all cores available.}
}
\description{
A rewrite of the pair-wise Spearman rank co-occurrence routine written by \href{https://github.com/germs-lab/FastCoOccur}{Jin Choi}. The routine has been adapted to integrate with the \code{\link[Rcpp]{Rcpp-package}} API.
}
\examples{
data(mock_phyloseq)
phylosmith:::co_occurrence_rho(mock_phyloseq, "day", 0.05)
}
\seealso{
\code{\link{bootstrap_rho}} \code{\link{phylosmith}}
}
\keyword{nonparametric}
