% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_treatments.R
\name{combine_treatments}
\alias{combine_treatments}
\title{Combine meta-data columns.}
\usage{
combine_treatments(phyloseq_obj, treatments, sep = '.')
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object created with the \link[=phyloseq]{phyloseq} package.}

\item{treatments}{A vector of column names or numbers in the \code{\link[phyloseq:sample_data]{sample_data}}.}

\item{sep}{Delimiter to separate the treatments.}
}
\description{
Combines multiple columns of a \code{\link[phyloseq]{phyloseq-class}} object \code{\link[phyloseq:sample_data]{sample_data}} into a single-variable column.
}
\examples{
data(mock_phyloseq)
combine_treatments(mock_phyloseq, treatments = c("treatment", "day"))@sam_data
combine_treatments(mock_phyloseq, treatments = c("treatment", "day"), sep = '_')@sam_data
}
\keyword{manip}
