% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate_treatments.R
\name{concatenate_treatments}
\alias{concatenate_treatments}
\alias{combine_treatments}
\title{Combine meta-data columns}
\usage{
concatenate_treatments(phyloseq_obj, treatments, sep = '.')
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object created with the \link[=phyloseq]{phyloseq} package.}

\item{treatments}{A vector of column names or numbers in the \code{\link[phyloseq:sample_data]{sample_data}}.}

\item{sep}{Delimiter to separate the treatments.}
}
\description{
This function takes a phyloseq object combines multiple columns in the meta-data into a single variable column.
}
\examples{
data(mock_phyloseq)
concatenate_treatments(mock_phyloseq, treatments = c("treatment", "day"))
concatenate_treatments(mock_phyloseq, treatments = c("treatment", "day"), sep = '_')
}
\keyword{manip}
