% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{conglomerate_samples}
\alias{conglomerate_samples}
\title{Merge samples based on common factor within sample_data.
Function from the phylosmith-package.}
\usage{
conglomerate_samples(phyloseq_obj, treatment, subset = NULL)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object. It
must contain \code{\link[phyloseq:sample_data]{sample_data()}}) with
information about each sample, and it must contain
\code{\link[phyloseq:tax_table]{tax_table()}}) with information about each
taxa/gene.}

\item{treatment}{Column name as a \code{string} or \code{numeric} in the
\code{\link[phyloseq:sample_data]{sample_data}}. This can be a vector of
multiple columns and they will be combined into a new column.}

\item{subset}{A factor within the \code{treatment}. This will remove any
samples that to not contain this factor. This can be a vector of multiple
factors to subset on.}
}
\value{
phyloseq-object
}
\description{
Inputs a phyloseq object and merges the samples that meet the
specified criteria into a single sample. This is meant for replicates, or
samples statistically proven to not be significantly different. This should
be used with caution as it may be a misleading representation of the data.
}
\examples{
conglomerate_samples(soil_column, treatment = c('Day', 'Matrix', 'Treatment'))
}
\seealso{
\code{\link[phyloseq:merge_samples]{merge_samples()}}
}
