% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Correlation}
\alias{Correlation}
\title{Co-occurrence calculation}
\usage{
Correlation(X, Y,
cor_coef_cutoff, p_cutoff, method, ncores)
}
\arguments{
\item{X}{An \code{otu_table} in the format from
\code{\link[phyloseq:otu_table]{phyloseq}}}

\item{Y}{An \code{otu_table} in the format from
\code{\link[phyloseq:otu_table]{phyloseq}}}

\item{cor_coef_cutoff}{\code{double} representing the minimum \code{rho-value}
accepted for the correlation to be returned.}

\item{p_cutoff}{\code{double} representing the maximum \code{p-value}
accepted for the correlation to be returned.}

\item{method}{Pearson, Spearman}

\item{ncores}{An \code{int} for how many cores to use to multithread the
calculations.}
}
\value{
A \code{data.frame} with treatment, otu_1, otu_2, rho, p values.
}
\description{
Calculate the pair-wise Spearman rank correlation.
}
\seealso{
\code{\link{co_occurrence}}
}
\author{
Schuyler D. Smith
}
