% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new2.R
\name{dendrogram_phyloseq}
\alias{dendrogram_phyloseq}
\title{Create a ggplot object of the dendrogram from a phyloseq object.
Function from the phylosmith-package.}
\usage{
dendrogram_phyloseq(phyloseq_obj, treatment, method = 'bray',
colors = 'default')
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object. It
must contain \code{\link[phyloseq:sample_data]{sample_data()}}) with
information about each sample, and it must contain
\code{\link[phyloseq:tax_table]{tax_table()}}) with information about each
taxa/gene.}

\item{treatment}{Column name as a string or number in the
\code{\link[phyloseq:sample_data]{sample_data}}. This can be a vector of
multiple columns and they will be combined into a new column.}

\item{method}{the distance measurement algorithm to use, match to
"euclidean", "manhattan", "canberra", "clark", "bray", "kulczynski",
"jaccard", "gower", "altGower", "morisita", "horn", "mountford", "raup",
"binomial", "chao", "cao" or "mahalanobis".}

\item{colors}{Name of a color set from the
\link[=RColorBrewer]{RColorBrewer} package or a vector palette of R-accepted
colors.}
}
\value{
ggplot-object
}
\description{
Inputs a \code{\link[phyloseq]{phyloseq-class}} object and
plots the dendrogram of samples colored by treatment.
}
\examples{
dendrogram_phyloseq(soil_column, c('Matrix', 'Treatment'))
}
