% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastCoOccur.R
\name{FastCoOccur}
\alias{FastCoOccur}
\alias{cooccurrence}
\alias{co_occurrence}
\title{an efficient c++ implementation of the pair-wise Spearman rank co-occurrence.}
\usage{
FastCoOccur(phyloseq_obj, treatment, p = 0.05)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object created with the \link[=phyloseq]{phyloseq} package.}

\item{treatment}{Column name or number, or vector of, in the \code{\link[phyloseq:sample_data]{sample_data}}.}

\item{p}{the p-value cutoff. all returned co-occurrences must have a p-value less than or equal to p.}
}
\description{
This function is a rewrite of the pair-wise Spearman rank co-occurence routine written by Jin Choi. I adapted the routine to integrate with the Rcpp API.
}
\examples{
data(mock_phyloseq)
FastCoOccur(mock_phyloseq, "day", 0.05)
}
\keyword{nonparametric}
