% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_phyloseq_generalists.R
\name{find_generalists}
\alias{find_generalists}
\title{Find Generalists}
\usage{
find_generalists(phyloseq_obj, frequency = 0, treatments = NULL,
subset = NULL, below = FALSE, drop_samples = FALSE)
}
\arguments{
\item{phyloseq_obj}{A phyloseq-class object created with the phyloseq package.}

\item{frequency}{The minimum proportion of samples the taxa is seen in.}

\item{treatments}{Name(s) or column number(s) in the sample_data(). Function then checks if taxa seen in frequency in each treatment.}

\item{subset}{If taxa not needed to be seen in all treatments, then can check only one particular treatment subset.}

\item{below}{Does frequency define the minimum or maximum, should the presence fall below frequency or not.}

\item{drop_samples}{Should the function remove samples that that are empty after removing taxa filtered by frequency.}
}
\description{
This function takes a phyloseq object and finds which taxa are seen in a given proportion of samples, either in the entire dataset, by treatment, or a particular treatment of interest.
}
\examples{
data(mock_phyloseq)
find_generalists(mock_phyloseq, frequency = 0.3)
find_generalists(mock_phyloseq, frequency = 0.3, treatments = "day")
find_generalists(mock_phyloseq, frequency = 0.3, treatments = 3)
find_generalists(mock_phyloseq, frequency = 0.3, treatments = c("day", "treatment"))
find_generalists(mock_phyloseq, frequency = 0.3, treatments = c("day", "treatment"), subset = "5-soil")
}
\keyword{frequency}
\keyword{generalists}
\keyword{phyloseq}
\keyword{phylosmith}
