% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_functions.R
\name{find_unique_taxa}
\alias{find_unique_taxa}
\title{Find unique taxa between treatments of a phyloseq object. Function from the phylosmith-package.}
\usage{
find_unique_taxa(phyloseq_obj, treatment, subset = NULL)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object. It must contain \code{\link[phyloseq:sample_data]{sample_data()}}) with information about each sample.}

\item{treatment}{Column name as a string or number in the \code{\link[phyloseq:sample_data]{sample_data}}. This can be a vector of multiple columns and they will be combined into a new column.}

\item{subset}{A factor within the \code{treatment}. This will remove any samples that to not contain this factor. This can be a vector of multiple factors to subset on.}
}
\description{
This function takes a \code{\link[phyloseq]{phyloseq-class}} object and returns taxa that are unique to a each treatment factor of the data.
}
\seealso{
\code{\link{find_common_taxa}}
}
\keyword{manip}
