% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{histogram_bootstrapped_rhos}
\alias{histogram_bootstrapped_rhos}
\title{Create a ggplot object of the distribution of rho values from bootstrap_rho(). Function from the phylosmith-package.}
\usage{
histogram_bootstrapped_rhos(bootstrapped_rhos, p = 0.05,
zeros = FALSE, x_scale = 0.25, colors = 'default')
}
\arguments{
\item{bootstrapped_rhos}{A \code{data.table} output from \code{\link[=bootstrap_rho]{bootstrap_rho}}.}

\item{p}{The significance threshold for setting cutoffs.}

\item{zeros}{In some cases either where a treatment is underpowered (few samples) there may be an extreme number of 0 values for rho, which can affect the scale an make the distribution curves look poor for treatments that are okay. The 0-rhos are still included in the calculations, but not displayed on the graph (\code{FALSE}).}

\item{x_breaks}{What intervals to set the ticks on the x-axis.}

\item{colors}{Name of a color set from the \link[=RColorBrewer]{RColorBrewer} package or a vector palete of R-accepted colors.}
}
\description{
Plots the output of \code{\link[=bootstrap_rho]{bootstrap_rho}} into a histogram with the distributions shown by treatment. This is a visualization tool to help show how the bootstrapping worked, and to see where the cutoffs lie.
}
\seealso{
\code{\link[=bootstrap_rho]{bootstrap_rho}}
}
