% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ia_lakes}
\alias{ia_lakes}
\title{IA Lake Data}
\format{
A \code{\link[phyloseq]{phyloseq-class}} object.
}
\source{
Simulation
}
\usage{
data(ia_lakes)
}
\description{
This data is from microbiomes associate with Lakes in 
Iowa, USA between 2018 and 2019.
}
\keyword{datasets}
