% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new2.R
\name{library_size}
\alias{library_size}
\title{Normalizes abundance data in an \code{otu_table} using library size.
Function from the phylosmith-package.}
\usage{
library_size(phyloseq_obj)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object. It
must contain \code{\link[phyloseq:sample_data]{sample_data()}} with
information about each sample, and it must contain
\code{\link[phyloseq:tax_table]{tax_table()}}) with information about each
taxa/gene.}
}
\value{
phyloseq-object
}
\description{
Transform abundance data to equal counts using library size and the geometric mean, 
i.e. proportional data.
}
\examples{
library_size(soil_column)
}
