% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_functions.R
\name{melt_phyloseq}
\alias{melt_phyloseq}
\title{Melt a phyloseq object into a data.table. Function from the phylosmith-package.}
\usage{
melt_phyloseq(phyloseq_obj)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object.}
}
\description{
Melts a \code{\link[phyloseq]{phyloseq-class}} object's otu_table, taxa_tale, and sample_data into a single data.table.
}
\seealso{
\code{\link[phyloseq:psmelt]{psmelt()}}
}
\keyword{manip}
