% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_functions.R
\name{merge_samples}
\alias{merge_samples}
\title{Merge samples based on common factor within sample_data. Function from the phylosmith-package.}
\usage{
merge_samples(phyloseq_obj, treatment, subset = NULL, merge_on = treatment)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object. It must contain \code{\link[phyloseq:sample_data]{sample_data()}}) with information about each sample.}

\item{treatment}{Column name as a string or number in the \code{\link[phyloseq:sample_data]{sample_data}}. This can be a vector of multiple columns and they will be combined into a new column.}

\item{subset}{A factor within the \code{treatment}. This will remove any samples that to not contain this factor. This can be a vector of multiple factors to subset on.}

\item{merge_on}{Defines which variable the data is merged according to. This needs to be a column name as a string or number in the \code{\link[phyloseq:sample_data]{sample_data}}. This can be a vector of multiple columns and they will be combined into a new column.}
}
\description{
Merges a \code{\link[phyloseq]{phyloseq-class}} object's samples that meet the specified criteria into a single sample. This is meant for replicates, or samples statistically proven to not be significantly different and should be used with caution as it may be a misleading representation of the data.
}
\seealso{
\code{\link[phyloseq:merge_samples]{merge_samples()}}
}
\keyword{manip}
