% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\name{network_phyloseq}
\alias{network_phyloseq}
\title{Create an object of the abundance barplots from a phyloseq object. Function from the phylosmith-package.}
\usage{
network_phyloseq(phyloseq_obj, treatment, subset = NULL, co_occurrence,
classification = 'none', colors = 'default')
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object created with the \link[=phyloseq]{phyloseq} package (must contain \code{\link[phyloseq:sample_data]{sample_data()}}).}

\item{treatment}{Column name or number, or vector of, in the \code{\link[phyloseq:sample_data]{sample_data}}.}

\item{subset}{If taxa not needed to be seen in all \code{treatment}, then will subset to treatments containing this string.}

\item{co_occurrence}{Co_Occurence table of the \code{phyloseq_obj}, computed using \code{\link{co_occurrence}}}

\item{classification}{Column name or number in the \code{\link[phyloseq:tax_table]{tax_table}} for node colors.}

\item{colors}{Name of a color set from the \link[=RColorBrewer]{RColorBrewer} package or a vector palete of R accepted colors.}
}
\description{
This function takes a \code{\link[phyloseq]{phyloseq-class}} object and creates barplots of taxa by treatment.
}
