% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{nmds_phyloseq}
\alias{nmds_phyloseq}
\title{Create a ggplot object of the NMDS from a phyloseq object.
Function from the phylosmith-package.}
\usage{
nmds_phyloseq(phyloseq_obj, treatment, method = 'bray', dimensions = 2, trymax = 100, 
circle = 0.95, labels = NULL, colors = 'default', verbose = TRUE)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object. It
must contain \code{\link[phyloseq:sample_data]{sample_data()}}) with
information about each sample, and it must contain
\code{\link[phyloseq:tax_table]{tax_table()}}) with information about each
taxa/gene.}

\item{treatment}{Column name as a string or number in the
\code{\link[phyloseq:sample_data]{sample_data}}. This can be a vector of
multiple columns and they will be combined into a new column.}

\item{method}{the distance measurement algorithm to use, match to
"euclidean", "manhattan", "canberra", "clark", "bray", "kulczynski",
"jaccard", "gower", "altGower", "morisita", "horn", "mountford", "raup",
"binomial", "chao", "cao" or "mahalanobis".}

\item{dimensions}{Number of dimensions. NB., the number of points n should 
be n > 2*k + 1, and preferably higher in global non-metric MDS, and still 
higher in local NMDS.}

\item{trymax}{Maximum numbers of random starts in search of stable 
solution. The iteration will stop when two convergent solutions were found or trymax was reached.}

\item{circle}{If TRUE, a \code{\link[ggplot2:stat_ellipse]{stat_ellipse}} around
each of the \code{treatment} factors (\code{TRUE}). If numeric between 0 and 1,
will add ellipse of confidence interval equal to value given (i.e. 0.95 produces
ellipses of 95\% confidence intervals)}

\item{labels}{Column name as a string or number in the
\code{\link[phyloseq:sample_data]{sample_data}} to use to place labels of
that factor instead of circle points.}

\item{colors}{Name of a color set from the
\link[=RColorBrewer]{RColorBrewer} package or a vector palette of R-accepted
colors.}

\item{verbose}{Whether or not to print the
\code{\link[vegan:metaMDS]{metaMDS}} stress convergence to console
(\code{TRUE}) or not (\code{FALSE}).}
}
\value{
ggplot-object
}
\description{
Inputs a \code{\link[phyloseq]{phyloseq-class}} object and
plots the NMDS of a treatment or set of treatments in space.
}
\examples{
nmds_phyloseq(soil_column, c('Matrix', 'Treatment'),
circle = TRUE, verbose = FALSE)
}
