% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_phyloseq_ggplot.R
\name{nmds_phyloseq_ggplot}
\alias{nmds_phyloseq_ggplot}
\title{Create a ggplot object of the NMDS from a phyloseq object. Function from the phylosmith-package.}
\usage{
nmds_phyloseq_ggplot(phyloseq_obj, treatment, circle = TRUE, labels = FALSE,
colors = 'default', verbose = TRUE)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object created with the \link[=phyloseq]{phyloseq} package (must contain \code{\link[phyloseq:sample_data]{sample_data()}}).}

\item{treatment}{Column name or number, or vector of, in the \code{\link[phyloseq:sample_data]{sample_data}}.}

\item{circle}{If TRUE, add elipses around each treatment.}

\item{labels}{Column name or number in the \code{\link[phyloseq:sample_data]{sample_data}} to use to place labels instead of points.}

\item{colors}{Name of a color set from the \link[=RColorBrewer]{RColorBrewer} package or a vector palete of R accepted colors.}

\item{verbose}{Whether or not to print the \code{\link[vegan:metaMDS]{metaMDS}} convergion to console or not.}
}
\description{
This function takes a \code{\link[phyloseq]{phyloseq-class}} object and plots the NMDS of a treatment or set of treatments.
}
