% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_filter.R
\name{order_phyloseq_metadata}
\alias{order_phyloseq_metadata}
\title{order_phyloseq_metadata}
\usage{
order_phyloseq_metadata(phyloseq_obj, treatment, order)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object created with the \link[=phyloseq]{phyloseq} package.}

\item{treatment}{Column name or number in the \code{\link[phyloseq:sample_data]{sample_data}}.}

\item{order}{The order of rank for the variables in \code{treatment} column.}
}
\description{
Reorders the levels of a metadata column in a \code{\link[phyloseq]{phyloseq-class}} object \code{\link[phyloseq:sample_data]{sample_data}}.
}
\keyword{manip}
