% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_functions.R
\name{order_treatment}
\alias{order_treatment}
\title{order_treatment}
\usage{
order_treatment(phyloseq_obj, treatment, order)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object. It must contain \code{\link[phyloseq:sample_data]{sample_data()}}) with information about each sample.}

\item{treatment}{Column name as a string or number in the \code{\link[phyloseq:sample_data]{sample_data}}.}

\item{order}{The order of factors in \code{treatment} column as a vector of strings.}
}
\description{
Reorders the levels of a metadata column in a \code{\link[phyloseq]{phyloseq-class}} object \code{\link[phyloseq:sample_data]{sample_data}}.
}
\keyword{manip}
