% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{pcoa_phyloseq}
\alias{pcoa_phyloseq}
\title{Create a ggplot object of the PCoA from a phyloseq object.
Function from the phylosmith-package.}
\usage{
pcoa_phyloseq(phyloseq_obj, treatment, x = 1, y = 2, method = 'bray',
circle = 0.95, colors = 'default', labels = NULL)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object. It
must contain \code{\link[phyloseq:sample_data]{sample_data()}}) with
information about each sample, and it must contain
\code{\link[phyloseq:tax_table]{tax_table()}}) with information about each
taxa/gene.}

\item{treatment}{Column name as a string or number in the
\code{\link[phyloseq:sample_data]{sample_data}}. This can be a vector of
multiple columns and they will be combined into a new column.}

\item{x}{the numerical principle compenent to use as the x-axis}

\item{y}{the numerical principle compenent to use as the y-axis}

\item{method}{the distance measurement algorithm to use, match to
"euclidean", "manhattan", "canberra", "clark", "bray", "kulczynski",
"jaccard", "gower", "altGower", "morisita", "horn", "mountford", "raup",
"binomial", "chao", "cao" or "mahalanobis".}

\item{circle}{If TRUE, a \code{\link[ggplot2:stat_ellipse]{stat_ellipse}} around
each of the \code{treatment} factors (\code{TRUE}). If numeric between 0 and 1,
will add ellipse of confidence interval equal to value given (i.e. 0.95 produces
ellipses of 95\% confidence intervals)}

\item{colors}{Name of a color set from the
\link[=RColorBrewer]{RColorBrewer} package or a vector palette of R-accepted
colors.}

\item{labels}{Column name as a string or number in the
\code{\link[phyloseq:sample_data]{sample_data}} to use to place labels of
that factor instead of circle points.}
}
\value{
ggplot-object
}
\description{
Inputs a \code{\link[phyloseq]{phyloseq-class}} object and
plots the PCoA of a treatment or set of treatments in space.
}
\examples{
pcoa_phyloseq(soil_column, c('Matrix', 'Treatment'),
circle = TRUE)
}
