% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_rho.R
\name{permute_rho}
\alias{permute_rho}
\title{permutes the pair-wise Spearman rank co-occurrence, to determine a significant rho-cutoff.}
\usage{
permute_rho(phyloseq_obj, treatment, replicates = 'independent', permutations = 10000, p = 0.05)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object created with the \link[=phyloseq]{phyloseq} package.}

\item{treatment}{Column name or number, or vector of, in the \code{\link[phyloseq:sample_data]{sample_data}}.}

\item{replicates}{Column name or number, or vector of, in the \code{\link[phyloseq:sample_data]{sample_data}} that indicates which samples are non-independent of each other.}

\item{permutations}{Number of iterations to compute.}

\item{p}{significance cut-off for the mean rho values.}
}
\description{
'
}
\examples{
data(mock_phyloseq)
permute_rho(mock_phyloseq, "day", permutations = 10)
}
\keyword{nonparametric}
