% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{permute_rho_Rcpp}
\alias{permute_rho_Rcpp}
\title{Co-occurrence rho calculations}
\usage{
permute_rho_Rcpp(count_matrix, permuted_matrix, method, ncores)
}
\arguments{
\item{count_matrix}{An \code{otu_table} in the format from
\code{\link[phyloseq:otu_table]{phyloseq}}}

\item{permuted_matrix}{An \code{otu_table} in the format from
\code{\link[phyloseq:otu_table]{phyloseq}}}

\item{method}{pearson, spearman}

\item{ncores}{An \code{int} for how many cores to use to multithread the
calculations.}
}
\value{
A \code{vector} with rho values for each pair-wise correlation.
}
\description{
Calculates the pair-wise Spearman rank correlation without
testing for significance.
}
\seealso{
\code{\link{permute_rho}}
}
\author{
Schuyler D. Smith
}
