% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{phylogeny_profile}
\alias{phylogeny_profile}
\title{Create a ggplot object of the phylogenic barplots from a phyloseq object.
Function from the phylosmith-package.}
\usage{
phylogeny_profile(phyloseq_obj, treatment = NULL, subset = NULL,
classification = NULL, merge = TRUE, relative_abundance = FALSE,
colors = 'default', grid = FALSE, treatment_labels = NULL,
sample_labels = NULL, classification_labels= NULL)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object. It
must contain \code{\link[phyloseq:sample_data]{sample_data()}}) with
information about each sample, and it must contain
\code{\link[phyloseq:tax_table]{tax_table()}}) with information about each
taxa/gene.}

\item{treatment}{Column name as a string or number in the
\code{\link[phyloseq:sample_data]{sample_data}}. This can be a vector of
multiple columns and they will be combined into a new column.}

\item{subset}{A factor within the \code{treatment}. This will remove any
samples that to not contain this factor. This can be a vector of multiple
factors to subset on.}

\item{classification}{Column name as a string or number in the
\code{\link[phyloseq:tax_table]{tax_table}} for the factor to use for node
colors.}

\item{merge}{if \code{FALSE}, will show separation of individuals within
each \code{classification}.}

\item{relative_abundance}{If \code{TRUE}, transforms the abundance data
into relative abundance by sample.}

\item{colors}{Name of a color set from the
\link[=RColorBrewer]{RColorBrewer} package or a vector palette of R-accepted
colors.}

\item{grid}{Wraps the sub-plots into a grid pattern rather than side-by-side.}

\item{treatment_labels}{a vector of names to be used as labels for
treatments/facets.}

\item{sample_labels}{a vector of names to be used as labels for Samples.}

\item{classification_labels}{a vector of names to be used as labels for the
taxonomic classifications.}
}
\value{
ggplot-object
}
\description{
Inputs a \code{\link[phyloseq]{phyloseq-class}} object and
creates phylogenic barplots.
}
\examples{
phylogeny_profile(soil_column, classification = 'Phylum',
treatment = c('Matrix', 'Treatment'), merge = TRUE,
relative_abundance = TRUE)
}
