% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_occurrence.R
\name{quantile_bootstrapped_rhos}
\alias{quantile_bootstrapped_rhos}
\title{Calculate quantiles for the bootstrapped rho values from the Spearman-rank co-occurrence. Function from the phylosmith-package.}
\usage{
quantile_bootstrapped_rhos(bootstrapped_rhos, p = 0.05, by_treatment = TRUE)
}
\arguments{
\item{bootstrapped_rhos}{A \code{data.table} output from \code{\link[=bootstrap_rho]{bootstrap_rho}}.}

\item{p}{The significance threshold for setting cutoffs.}

\item{by_treatment}{Whether to find the rho cutoffs for each treatment individually or for the entire experiment. Suggested to do by treatment first, to see if there is any treatments that are outliers.}
}
\description{
Calculate quantiles for the bootstrapped rho values from the Spearman-rank co-occurrence.
}
\seealso{
\code{\link[=bootstrap_rho]{bootstrap_rho}}
}
