% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{set_sample_order}
\alias{set_sample_order}
\title{Re-orders the samples of a phyloseq object.
Function from the phylosmith-package.}
\usage{
set_sample_order(phyloseq_obj, sort_on)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object. It
must contain \code{\link[phyloseq:sample_data]{sample_data()}}) with
information about each sample, and it must contain
\code{\link[phyloseq:tax_table]{tax_table()}}) with information about each
taxa/gene.}

\item{sort_on}{Column name as a \code{string} or \code{numeric} in the
\code{\link[phyloseq:sample_data]{sample_data}}, or vector of sample names
or indices in particular order.}
}
\value{
phyloseq object
}
\description{
Inputs a \code{\link[phyloseq]{phyloseq-class}} object and changes the
order of sample index either based on the metadata, or a given order. If
metadata columns are used, they will take the factor order.
}
\examples{
set_sample_order(soil_column, c('Matrix', 'Treatment'))
}
