% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{set_treatment_levels}
\alias{set_treatment_levels}
\title{set_treatment_levels}
\usage{
set_treatment_levels(phyloseq_obj, treatment, order)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object. It
must contain \code{\link[phyloseq:sample_data]{sample_data()}}) with
information about each sample, and it must contain
\code{\link[phyloseq:tax_table]{tax_table()}}) with information about each
taxa/gene.}

\item{treatment}{Column name as a \code{string} or \code{numeric} in the
\code{\link[phyloseq:sample_data]{sample_data}}.}

\item{order}{The order of factors in \code{treatment} column as a vector of
strings. If assigned "numeric" will set ascending numerical order.}
}
\value{
phyloseq-object
}
\description{
Reorders the levels of a metadata factor column in a
\code{\link[phyloseq]{phyloseq-class}} object
\code{\link[phyloseq:sample_data]{sample_data}}.
}
\examples{
set_treatment_levels(soil_column, treatment = 'Matrix',
order = c('Manure', 'Soil', 'Effluent'))
set_treatment_levels(soil_column, 'Day', 'numeric')
}
