% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{soil_column}
\alias{soil_column}
\title{Soil Column 16S Data - OTUs}
\format{A \code{\link[phyloseq]{phyloseq-class}} object.}
\source{
Simulation
}
\usage{
data(soil_column)
}
\description{
Soil Column 16S Data - OTUs
}
\examples{
data(soil_column)
otu_table(soil_column)
sample_data(soil_column)
}
\references{
\url{https://www.frontiersin.org/articles/10.3389/fmicb.2018.03197/full}
}
\keyword{datasets}
