% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{taxa_abundance_bars}
\alias{taxa_abundance_bars}
\title{Create a ggplot object of the abundance barplots from a phyloseq object.
Function from the phylosmith-package.}
\usage{
taxa_abundance_bars(phyloseq_obj, treatment = NULL,
classification = NULL, subset = NULL, transformation = 'none',
colors = 'default', wrap_by = NULL, treatment_labels = NULL,
sample_labels = NULL, classification_labels= NULL)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object. It
must contain \code{\link[phyloseq:sample_data]{sample_data()}}) with
information about each sample, and it must contain
\code{\link[phyloseq:tax_table]{tax_table()}}) with information about each
taxa/gene.}

\item{treatment}{Column name as a string or number in the
\code{\link[phyloseq:sample_data]{sample_data}}. This can be a vector of
multiple columns and they will be combined into a new column.}

\item{subset}{A factor within the \code{treatment}. This will remove any
samples that to not contain this factor. This can be a vector of multiple
factors to subset on.}

\item{classification}{Column name as a string or number in the
\code{\link[phyloseq:tax_table]{tax_table}} for the factor to use for node
colors.}

\item{transformation}{Transformation to be used on the data. "none",
"mean", "median", "sd", "log", "log10"}

\item{colors}{Name of a color set from the
\link[=RColorBrewer]{RColorBrewer} package or a vector palette of R-accepted
colors.}

\item{wrap_by}{Column name as a string or number in the
phyloseq \code{\link[phyloseq:sample_data]{sample_data}} to facet the ggplot
figure by.}

\item{treatment_labels}{a vector of names to be used as labels for
treatments/facets.}

\item{sample_labels}{a vector of names to be used as labels for Samples.}

\item{classification_labels}{a vector of names to be used as labels for the
taxonomic classifications.}
}
\value{
ggplot-object
}
\description{
Inputs a \code{\link[phyloseq]{phyloseq-class}} object and
creates barplots of taxa by treatment.
}
\examples{
taxa_abundance_bars(
taxa_filter(soil_column, frequency = 0.8),
classification = 'Phylum', treatment = c('Matrix', 'Treatment'),
subset = 'Unamended', transformation = 'mean')
}
