% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_phyloseq_ggplot.R
\name{taxa_abundance_bars_ggplot}
\alias{taxa_abundance_bars_ggplot}
\title{Create a ggplot object of the abundance barplots from a phyloseq object. Function from the phylosmith-package.}
\usage{
taxa_abundance_bars_ggplot(phyloseq_obj, classification = 'none', treatment,
subset = NULL, transformation = 'none', colors = 'default')
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object created with the \link[=phyloseq]{phyloseq} package (must contain \code{\link[phyloseq:sample_data]{sample_data()}}).}

\item{classification}{Column name or number in the \code{\link[phyloseq:tax_table]{tax_table}} for the x-axis.}

\item{treatment}{Column name or number, or vector of, in the \code{\link[phyloseq:sample_data]{sample_data}}.}

\item{subset}{If taxa not needed to be seen in all \code{treatment}, then will subset to treatments containing this string.}

\item{transformation}{Transforms the data. "none", "mean", "median", "sd", "log", "log10"}

\item{colors}{Name of a color set from the \link[=RColorBrewer]{RColorBrewer} package or a vector palete of R accepted colors.}
}
\description{
This function takes a \code{\link[phyloseq]{phyloseq-class}} object and creates barplots of taxa by treatment.
}
