% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analytics.R
\name{taxa_core}
\alias{taxa_core}
\title{Filter taxa in phyloseq-object to only include core taxa.
Function from the phylosmith-package.}
\usage{
taxa_core(phyloseq_obj, treatment = NULL, subset = NULL,
frequency = 0.5, abundance_threshold = 0.01)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object. It
must contain \code{\link[phyloseq:sample_data]{sample_data()}}) with
information about each sample, and it must contain
\code{\link[phyloseq:tax_table]{tax_table()}}) with information about each
taxa/gene.}

\item{treatment}{Column name as a \code{string} or \code{numeric} in the
\code{\link[phyloseq:sample_data]{sample_data}}. This can be a vector of
multiple columns and they will be combined into a new column.}

\item{subset}{A factor within the \code{treatment}. This will remove any
samples that to not contain this factor. This can be a vector of multiple
factors to subset on.}

\item{frequency}{The proportion of samples the taxa is found in.}

\item{abundance_threshold}{The minimum relative abundance the taxa is found
in for each sample.}
}
\value{
phyloseq-object
}
\description{
Inputs a phyloseq object and finds which taxa are seen in a
given proportion of samples at a minimum relative abundance, either in the
entire dataset, by treatment, or a particular treatment of interest.
}
\examples{
taxa_core(soil_column, frequency = 0.2, abundance_threshold = 0.01)
}
