% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{taxa_core_graph}
\alias{taxa_core_graph}
\title{Create graph of the core taxa seen in phyloseq-object over a range of
abundance and smaple-frequency values.  Function from the phylosmith-package.}
\usage{
taxa_core_graph(phyloseq_obj, treatment = NULL, subset = NULL,
frequencies = seq(0.1, 1, 0.1), abundance_thresholds = seq(0.01, 1, 0.01),
colors = 'default',
treatment_labels = NULL, sample_labels = NULL, classification_labels= NULL)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object. It
must contain \code{\link[phyloseq:sample_data]{sample_data()}}) with
information about each sample, and it must contain
\code{\link[phyloseq:tax_table]{tax_table()}}) with information about each
taxa/gene.}

\item{treatment}{Column name as a \code{string} or \code{numeric} in the
\code{\link[phyloseq:sample_data]{sample_data}}. This can be a vector of
multiple columns and they will be combined into a new column.}

\item{subset}{A factor within the \code{treatment}. This will remove any
samples that to not contain this factor. This can be a vector of multiple
factors to subset on.}

\item{frequencies}{The range of proportions of samples the taxa are found in.}

\item{abundance_thresholds}{The range of minimum relative abundances the taxa are found
in for each sample.}

\item{colors}{Name of a color set from the
\link[=RColorBrewer]{RColorBrewer} package or a vector palette of R-accepted
colors.}

\item{treatment_labels}{a vector of names to be used as labels for
treatments/facets.}

\item{sample_labels}{a vector of names to be used as labels for Samples.}

\item{classification_labels}{a vector of names to be used as labels for the
taxonomic classifications.}
}
\value{
phyloseq-object
}
\description{
Inputs a phyloseq object and finds which taxa are seen in a
given proportion of samples at a minimum relative abundance, either in the
entire dataset, or by treatment, over a range of values. Then draws the
distribution.
}
\examples{
taxa_core_graph(soil_column)
}
