% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{taxa_extract}
\alias{taxa_extract}
\title{Extracts specific taxa from all samples based on a given classification level.
Function from the phylosmith-package.}
\usage{
taxa_extract(phyloseq_obj, taxa_to_extract, classification=NULL)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object. It
must contain \code{\link[phyloseq:sample_data]{sample_data()}} with
information about each sample, and it must contain
\code{\link[phyloseq:tax_table]{tax_table()}}) with information about each
taxa/gene.}

\item{taxa_to_extract}{A vector of the classification names of the taxon to
be extracted. If you are using ASVs and want to extract specific sequences, this
should be set OTU or NULL. By default, it will search for the names in all
taxanomic ranks}

\item{classification}{Column name as a \code{string} or \code{numeric} in
the \code{\link[phyloseq:tax_table]{tax_table}} for the factor to
conglomerate by.}
}
\value{
phyloseq-object
}
\description{
Extracts taxa from phyloseq objects based on taxanomic names.
}
\examples{
taxa_extract(soil_column, 'Firmicutes', 'Phylum')
}
\seealso{
\code{\link[phyloseq:tax_glom]{tax_glom()}}
}
