% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_filter.R
\name{taxa_filter}
\alias{taxa_filter}
\title{Filter taxa based on proportion of samples they are observed in. Function from the phylosmith-package.}
\usage{
taxa_filter(phyloseq_obj, treatment = NULL, frequency = 0,
subset = NULL, below = FALSE, drop_samples = FALSE)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object. It must contain \code{\link[phyloseq:sample_data]{sample_data()}}) with information about each sample, and it must contain \code{\link[phyloseq:tax_table]{tax_table()}}) with information about each taxa/gene.}

\item{treatment}{Column name as a string or number in the \code{\link[phyloseq:sample_data]{sample_data}}. This can be a vector of multiple columns and they will be combined into a new column.}

\item{frequency}{The proportion of samples the taxa is found in.}

\item{subset}{A factor within the \code{treatment}. This will remove any samples that to not contain this factor. This can be a vector of multiple factors to subset on.}

\item{below}{Does frequency define the minimum (\code{FALSE}) or maximum (\code{TRUE}) proportion of samples the taxa is found in.}

\item{drop_samples}{Should the function remove samples that that are empty after removing taxa filtered by frequency (\code{TRUE}).}
}
\description{
This function takes a phyloseq object and finds which taxa are seen in a given proportion of samples, either in the entire dataset, by treatment, or a particular treatment of interest.
}
\keyword{manip}
