% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_parsing.R
\name{taxa_proportions}
\alias{taxa_proportions}
\title{Compute proportions for taxa.}
\usage{
taxa_proportions(phyloseq_obj, classification, treatment = NA)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object. it 
must contain \code{\link[phyloseq:tax_table]{tax_table()}}) with 
information about each taxa/gene.}

\item{classification}{Column name as a \code{string} or \code{numeric} in 
the \code{\link[phyloseq:tax_table]{tax_table}} for the prportions to be 
reported on.}

\item{treatment}{Column name as a \code{string} or \code{numeric} in the 
\code{\link[phyloseq:sample_data]{sample_data}}. This can be a vector of 
multiple columns and they will be combined into a new column. If \code{NA}, 
then proprtions will be reported for the entire dataset. If set to "sample" 
it will repoort proportions by sample.}
}
\description{
Computes the proportion of a taxa classification. Function from the 
phylosmith-package.
}
