% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{taxa_prune}
\alias{taxa_prune}
\title{Remove specific taxa from all samples based on a given classification level.
Function from the phylosmith-package.}
\usage{
taxa_prune(phyloseq_obj, taxa_to_remove,
classification=NULL, na.rm=FALSE)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object. It
must contain \code{\link[phyloseq:sample_data]{sample_data()}} with
information about each sample, and it must contain
\code{\link[phyloseq:tax_table]{tax_table()}}) with information about each
taxa/gene.}

\item{taxa_to_remove}{A vector of the classification names of the taxon to
be removed. If you are using ASVs and want to remove specific sequences, this
should be set OTU or NULL. By default, it will search for the names in all
taxanomic ranks}

\item{classification}{Column name as a \code{string} or \code{numeric} in
the \code{\link[phyloseq:tax_table]{tax_table}} for the factor to
conglomerate by.}

\item{na.rm}{if TRUE, and classification is specified, will remove taxon at
the classificaiton level that have NA values.}
}
\value{
phyloseq-object
}
\description{
Prunes taxa from phyloseq objects based on taxanomic names.
}
\examples{
taxa_prune(soil_column, 'Firmicutes', 'Phylum')
}
\seealso{
\code{\link[phyloseq:tax_glom]{tax_glom()}}
}
