% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlations.R
\name{variable_correlation}
\alias{variable_correlation}
\title{Computes the correlation of numerical variables with taxa
Function from the phylosmith-package.}
\usage{
variable_correlation(phyloseq_obj, variables, treatment = NULL,
 subset = NULL, classification = NULL, method = 'spearman', cores = 1)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object.}

\item{variables}{Numericla factors within the in the
\code{\link[phyloseq:sample_data]{sample_data}} to correlate with the
abundance data.}

\item{treatment}{Column name as a \code{string} or \code{numeric} in the
\code{\link[phyloseq:sample_data]{sample_data}}. This can be a vector of
multiple columns and they will be combined into a new column.}

\item{subset}{A factor within the \code{treatment}. This will remove any
samples that to not contain this factor. This can be a vector of multiple
factors to subset on.}

\item{classification}{Column name as a \code{string} or \code{numeric} in
the \code{\link[phyloseq:tax_table]{tax_table}} for the factor to
conglomerate by.}

\item{method}{Which correlation method to calculate, "pearson", "spearman".}

\item{cores}{\code{numeric} Number of CPU cores to use for the pair-wise
permutations. Default (0) uses max cores available. Parallelization not
available for systems running MacOS without openMP configuration.}
}
\value{
data.table
}
\description{
Computes the correlation of numerical variables with taxa
}
\examples{
variable_correlation(soil_column, variables = 'Day',
treatment = c('Matrix', 'Treatment'), subset = 'Amended',
classification = 'Phylum', method = 'spearman', cores = 1)
}
\seealso{
\code{\link{permute_rho}} \code{\link{phylosmith}}
}
\keyword{nonparametric}
