% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{variable_correlation_heatmap}
\alias{variable_correlation_heatmap}
\title{Computes the correlation of numerical variables with taxa
Function from the phylosmith-package.}
\usage{
variable_correlation_heatmap(phyloseq_obj, treatment = NULL,
subset = NULL, classification = NULL, variables,
method = 'spearman', limits = c(-0.8, 0.8),
colors = 'default', significance_color = 'white', cores = 1,
treatment_labels = NULL, sample_labels = NULL, classification_labels= NULL)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object.}

\item{treatment}{Column name as a \code{string} or \code{numeric} in the
\code{\link[phyloseq:sample_data]{sample_data}}. This can be a vector of
multiple columns and they will be combined into a new column.}

\item{subset}{A factor within the \code{treatment}. This will remove any
samples that to not contain this factor. This can be a vector of multiple
factors to subset on.}

\item{classification}{Column name as a \code{string} or \code{numeric} in
the \code{\link[phyloseq:tax_table]{tax_table}} for the factor to
conglomerate by.}

\item{variables}{Numerical factors within the in the
\code{\link[phyloseq:sample_data]{sample_data}} to correlate with the
abundance data.}

\item{method}{Which correlation method to calculate, "pearson", "spearman".}

\item{limits}{The range for the legend, smaller limits will accentuate smaller
correlations.}

\item{colors}{the palette to use for the heatmap, default is viridis.}

\item{significance_color}{the color to use for the significance stars.}

\item{cores}{\code{numeric} Number of CPU cores to use for the pair-wise
permutations. Default (1), (0) uses max cores available. Parallelization not
available for systems running MacOS without openMP configuration.}

\item{treatment_labels}{a vector of names to be used as labels for
treatments/facets.}

\item{sample_labels}{a vector of names to be used as labels for Samples.}

\item{classification_labels}{a vector of names to be used as labels for the
taxonomic classifications.}
}
\value{
data.table
}
\description{
Computes the correlation of numerical variables with taxa
}
\examples{
variable_correlation_heatmap(soil_column, variables = 'Day',
treatment = c('Matrix', 'Treatment'), subset = 'Amended',
classification = 'Phylum', method = 'spearman', cores = 1,
colors = c("#2C7BB6", "white", "#D7191C"),
significance_color = 'black')
}
\seealso{
\code{\link{permute_rho}} \code{\link{phylosmith}}
}
\keyword{nonparametric}
