% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{variable_correlation_network}
\alias{variable_correlation_network}
\title{Create a node network ggplot object of the correlation of taxa and sample variables
from a phyloseq object. Function from the phylosmith-package.}
\usage{
variable_correlation_network(phyloseq_obj, variables, classification = NULL,
treatment = NULL, subset = NULL, correlation_table = NULL, method = 'spearman',
rho_threshold = c(-0.01, 0.01), p_threshold = 0.05, colors = 'default',
negative_positive_colors = c('pink1', 'gray22'))
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object. It
must contain \code{\link[phyloseq:sample_data]{sample_data()}}) with
information about each sample, and it must contain
\code{\link[phyloseq:tax_table]{tax_table()}}) with information about each
taxa/gene.}

\item{variables}{Numerical factors within the in the
\code{\link[phyloseq:sample_data]{sample_data}} to correlate with the
abundance data.}

\item{classification}{Column name as a string or number in the
\code{\link[phyloseq:tax_table]{tax_table}} for the factor to use for node
colors.}

\item{treatment}{Column name as a string or number in the
\code{\link[phyloseq:sample_data]{sample_data}}. This can be a vector of
multiple columns and they will be combined into a new column.}

\item{subset}{A factor within the \code{treatment}. This will remove any
samples that to not contain this factor. This can be a vector of multiple
factors to subset on.}

\item{correlation_table}{Table of the correlation of taxa/variables in the
\code{phyloseq_obj}, computed using \code{\link{variable_correlation}}. If no
table is given, it will be computed with the \code{phyloseq_obj}, using the
given \code{treatment} and \code{p} = 0.05.}

\item{method}{Which correlation method to calculate, "pearson", "spearman", "kendall.}

\item{rho_threshold}{Cutoffs to use to subset the `correlation_table` by correlation values.}

\item{p_threshold}{Cutoffs to use to subset the `correlation_table` by singnificance values.}

\item{colors}{Name of a color set from the #' \link[=RColorBrewer]{RColorBrewer}
package or a vector palette of R accepted colors.}

\item{negative_positive_colors}{colors to use for the edges to represent negative and
positive correlations.}
}
\value{
ggplot-object
}
\description{
Inputs a \code{\link[phyloseq]{phyloseq-class}} object and
creates a network from the co-occurrence. The co-occurrence can either be
input, or it will be calculated with the Spearman-rank correlation. Also,
the layout of the graph can be given as an argument as well for reproducibility.
}
