% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{abundance_heatmap}
\alias{abundance_heatmap}
\title{Create a heatmap of the abundance table from a
phyloseq object. Function from the phylosmith-package.}
\usage{
abundance_heatmap(phyloseq_obj, treatment, subset = NULL,
classification = NULL, transformation = 'none', colors = 'default',
treatment_labels = NULL, sample_labels = NULL, classification_labels= NULL)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object. It
must contain \code{\link[phyloseq:sample_data]{sample_data()}}) with
information about each sample, and it must contain.
\code{\link[phyloseq:tax_table]{tax_table()}}) with information about
each taxa/gene.}

\item{treatment}{Column name as a string or number in the
\code{\link[phyloseq:sample_data]{sample_data}}. This can be a vector of
multiple columns and they will be combined into a new column.}

\item{subset}{A factor within the \code{treatment}. This will remove any
samples that to not contain this factor. This can be a vector of multiple
factors to subset on.}

\item{classification}{Column name as a string or number in the
\code{\link[phyloseq:tax_table]{tax_table}} for the factor.}

\item{transformation}{Transformation to be used on the data. "none",
"relative_abundance", "log", "log10", "log1p", "log2", "asn", "atanh",
"boxcox", "exp", "identity", "logit", "probability", "probit",
"reciprocal", "reverse" and "sqrt"}

\item{colors}{Name of a color set from the
\link[=RColorBrewer]{RColorBrewer} package or a vector palette of R-accepted
colors.}

\item{treatment_labels}{a vector of names to be used as labels for
treatments/facets.}

\item{sample_labels}{a vector of names to be used as labels for Samples.}

\item{classification_labels}{a vector of names to be used as labels for the
taxonomic classifications.}
}
\value{
ggplot-object
}
\description{
Takes a \code{\link[phyloseq]{phyloseq-class}} object as input and
creates a ggplot-heatmap of the abundances across samples.
The default color choice is the viridis palette, which is supposed to
be both aesthetic for normal and color-blind viewers.
}
\examples{
abundance_heatmap(soil_column, classification = 'Phylum',
treatment = c('Matrix', 'Treatment'), transformation = 'log')
}
