% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_rho.R
\name{bootstrap_rho}
\alias{bootstrap_rho}
\title{Bootstraps the pair-wise Spearman rank co-occurrence, to determine a significant rho-cutoff. phylosmith}
\usage{
bootstrap_rho(phyloseq_obj, treatment, replicates = 'independent', permutations = 10000, p = 0.05)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object created with the \link[=phyloseq]{phyloseq} package.}

\item{treatment}{Column name or number, or vector of, in the \code{\link[phyloseq:sample_data]{sample_data}}.}

\item{replicates}{Column name or number, or vector of, in the \code{\link[phyloseq:sample_data]{sample_data}} that indicates which samples are non-independent of each other.}

\item{permutations}{Number of iterations to compute.}

\item{p}{Significance cut-off for the mean rho values. If set to 0 (default) returns a vector of all permuted rho values.}
}
\description{
Bootstraps the pair-wise Spearman rank co-occurrence, to determine a significant rho-cutoff.
}
\examples{
data(mock_phyloseq)
bootstrap_rho(mock_phyloseq, "day", permutations = 10)
}
\seealso{
\code{\link{FastCoOccur}}
}
\keyword{nonparametric}
