% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_export.R
\name{check_numeric_classification}
\alias{check_numeric_classification}
\title{Converts numeric values to column names in tax_table}
\usage{
check_numeric_classification(phyloseq_obj, ...)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object. It
must contain \code{\link[phyloseq:sample_data]{sample_data()}}) with
information about each sample, and it must contain
\code{\link[phyloseq:tax_table]{tax_table()}}) with information about each
taxa/gene.}

\item{...}{Column name as a \code{string} or \code{numeric} in the
\code{\link[phyloseq:tax_table]{tax_table}} for the factor to conglomerate
by.}
}
\value{
string
}
\description{
Converts numeric values to column names in tax_table.
}
\examples{
check_numeric_classification(soil_column, 2)
}
