% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_export.R
\name{check_numeric_treatment}
\alias{check_numeric_treatment}
\title{Converts numeric values to column names in sample_data.}
\usage{
check_numeric_treatment(phyloseq_obj, ...)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object. It
must contain \code{\link[phyloseq:sample_data]{sample_data()}}) with
information about each sample, and it must contain
\code{\link[phyloseq:tax_table]{tax_table()}}) with information about each
taxa/gene.}

\item{...}{Column name as a \code{string} or \code{numeric} in the
\code{\link[phyloseq:sample_data]{sample_data}}. This can be a vector of
multiple columns and they will be combined into a new column.}
}
\value{
string
}
\description{
Converts numeric values to column names in sample_data.
}
\examples{
check_numeric_treatment <- function(phyloseq_obj, ...){
treatments <- list(...)
treatments <- unlist(sapply(treatments, FUN = function(treatment){
  if(is.numeric(treatment)){
    return(colnames(access(phyloseq_obj, 'sam_data')[,treatment]))
  } else {return(treatment)}
}))
return(treatments)
}
check_numeric_treatment(soil_column, 2)
}
