% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analytics.R
\name{common_taxa}
\alias{common_taxa}
\title{Find taxa shared between treatments of a phyloseq object.
Function from the phylosmith-package.}
\usage{
common_taxa(phyloseq_obj, treatment = NULL, subset = NULL, n = 'all')
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object. It
must contain \code{\link[phyloseq:sample_data]{sample_data()}}) with
information about each sample, and it must contain
\code{\link[phyloseq:tax_table]{tax_table()}}) with information about each
taxa/gene.}

\item{treatment}{Column name as a \code{string} or \code{numeric} in the
\code{\link[phyloseq:sample_data]{sample_data}}. This can be a vector of
multiple columns and they will be combined into a new column.}

\item{subset}{A factor within the \code{treatment}. This will remove any
samples that to not contain this factor. This can be a vector of multiple
factors to subset on.}

\item{n}{Number of treatment groups that need to share the taxa to be
considered a common taxa.}
}
\value{
vector
}
\description{
Inputs a \code{\link[phyloseq]{phyloseq-class}} object and finds which taxa
are shared between all of the specified treatments from data in the
\code{\link[phyloseq:sample_data]{sample_data()}}), or every sample in the
dataset.
}
\examples{
common_taxa(soil_column, treatment = 'Treatment',
subset = 'Amended', n = 'all')
}
\seealso{
\code{\link{unique_taxa}}
}
