% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_functions.R
\name{conglomerate_taxa}
\alias{conglomerate_taxa}
\title{Conglomerate taxa by sample on a given classification level. Function from the phylosmith-package.}
\usage{
conglomerate_taxa(phyloseq_obj, classification, taxa_are_ordered = TRUE)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object. It must contain \code{\link[phyloseq:tax_table]{tax_table()}}) with information about each taxa/gene.}

\item{classification}{Column name as a string or number in the \code{\link[phyloseq:tax_table]{tax_table}} for the factor to conglomerate by.}

\item{taxa_are_ordered}{Whether the order of factors in the tax_table represent a decreasing heirarchy (TRUE) or are independant (FALSE). If FALSE, will only return the factor given by \code{classification}.}
}
\description{
Conglomerate taxa by sample on a given classification level from the tax_table. Similar to \code{phyloseq::tax_glom}
}
\seealso{
\code{\link[phyloseq:tax_glom]{tax_glom()}}
}
\keyword{manip}
