% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{conglomerate_taxa}
\alias{conglomerate_taxa}
\title{Conglomerate taxa by sample on a given classification level.
Function from the phylosmith-package.}
\usage{
conglomerate_taxa(phyloseq_obj, classification, hierarchical = TRUE,
use_taxonomic_names = TRUE)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object. It
must contain \code{\link[phyloseq:sample_data]{sample_data()}} with
information about each sample, and it must contain
\code{\link[phyloseq:tax_table]{tax_table()}}) with information about each
taxa/gene.}

\item{classification}{Column name as a \code{string} or \code{numeric} in
the \code{\link[phyloseq:tax_table]{tax_table}} for the factor to
conglomerate by.}

\item{hierarchical}{Whether the order of factors in the tax_table represent
a decreasing hierarchy (TRUE) or are independent (FALSE). If FALSE, will
only return the factor given by \code{classification}.}

\item{use_taxonomic_names}{If (TRUE), will use the hierarchical taxonomic
name from Domain to classification level. If (FALSE) will uses the
classification level with a numerical assignment.}
}
\value{
phyloseq-object
}
\description{
Conglomerate taxa by sample on a given classification level from the
tax_table.
}
\examples{
conglomerate_taxa(soil_column, 'Phylum')
}
\seealso{
\code{\link[phyloseq:tax_glom]{tax_glom()}}
}
