% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineternal.R
\name{create_palette}
\alias{create_palette}
\title{Creates color palettes for figures.}
\usage{
create_palette(color_count, colors = 'default')
}
\arguments{
\item{color_count}{Number of colors to choose for palette.}

\item{colors}{Name of a color set from the
\link[=RColorBrewer]{RColorBrewer} package or a vector palete of R-accepted
colors.}
}
\value{
palette
}
\description{
Creates color palettes for figures using the the RColorBrewer, or the default
color palette I made. The first 8 colors of the palette are from B. Wong's 2011
publication. The additional colors I added and may change as time goes by.
}
\examples{
#create_palette(8, 'Dark2')
}
