% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curate_cooccurrence.R
\name{curate_cooccurrence}
\alias{curate_cooccurrence}
\title{Curate co-occurrence data. Function from the phylosmith-package.}
\usage{
curate_cooccurrence(cooccurrence_table, taxa_of_interest, number_of_treatments = 1)
}
\arguments{
\item{cooccurrence_table}{co-occurrence table generated with \code{\link{FastCoOccur}}, or formatted in the same way.}

\item{taxa_of_interest}{a list or vector of taxa names, like those generated with \code{\link{find_unique_taxa}}.}

\item{number_of_treatments}{how many treatments should the taxa of interest be seen in? require \code{integer} or 'all' (default = 1).}
}
\description{
Used to curate a co-occurrence table from the \code{\link{FastCoOccur}} function. Takes a list of taxa and finds all pairs containing any of those taxa.
}
\examples{
data(mock_phyloseq)
}
\keyword{manip}
