% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curate_cooccurrence.R
\name{curate_cooccurrence}
\alias{curate_cooccurrence}
\title{Curate Co-Occurrence}
\usage{
curate_cooccurrence(cooccurrence_table, taxa_of_interest, number_of_treatments = 'all')
}
\arguments{
\item{cooccurrence_table}{co-occurrence table generated with FastCoOccur().}

\item{taxa_of_interest}{a list or vector of taxa names.}

\item{number_of_treatments}{how many treatments should the taxa of interest be seen in? require `int` or 'all' (default).}
}
\description{
This is used to to curate a co-occurrence table from the FastCoOccur function. It will take a list of taxa, find all pairs with those taxa.
}
\examples{
data(mock_phyloseq)
}
\keyword{cooccurrence}
\keyword{phyloseq}
\keyword{phylosmith}
\keyword{taxa}
