% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastCoOccur.R
\name{FastCoOccur}
\alias{FastCoOccur}
\title{FastCoOccur}
\usage{
FastCoOccur(phyloseq_obj, treatment, p)
}
\arguments{
\item{phyloseq_obj}{A phyloseq-class object created with the phyloseq package.}

\item{treatment}{the column name or number of the treatment to be comapred.}

\item{p}{the p-value cutoff. all returned co-occurrences must have a p-value less than or equal to p.}
}
\description{
This function is a rewrite of the pair-wise Spearman rank co-occurence routine written by Jin Choi. I adapted the routine to integrate with the Rcpp API.
}
\examples{
data(mock_phyloseq)
FastCoOccur(mock_phyloseq, "day", 0.05)
}
\keyword{Spearman}
\keyword{co-occurrence}
\keyword{coocurrence}
\keyword{phylosmith}
