% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastCoOccur.R
\name{FastCoOccur}
\alias{FastCoOccur}
\title{pair-wise Spearman rank co-occurrence, written in efficient c++ code. Function from the phylosmith-package.}
\usage{
FastCoOccur(phyloseq_obj, treatment, p = 0.05, cores = NULL)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object created with the \link[=phyloseq]{phyloseq} package.}

\item{treatment}{Column name or number, or vector of, in the \code{\link[phyloseq:sample_data]{sample_data}}.}

\item{p}{the p-value cutoff. all returned co-occurrences must have a p-value less than or equal to p.}

\item{cores}{Number of CPU cores to use for the pair-wise permutations. Default uses all cores available.}
}
\description{
A rewrite of the pair-wise Spearman rank co-occurence routine written by \href{https://github.com/germs-lab/FastCoOccur}{Jin Choi}. The routine has been adapted to integrate with the \link[=Rcpp]{Rcpp} API.
}
\examples{
data(mock_phyloseq)
FastCoOccur(mock_phyloseq, "day", 0.05)
}
\seealso{
\code{\link{phylosmith}}
}
\keyword{nonparametric}
