% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_unique_taxa.R
\name{find_common_taxa}
\alias{find_common_taxa}
\title{Find taxa shared between treatments of a phyloseq object. Function from the phylosmith-package.}
\usage{
find_common_taxa(phyloseq_obj, treatment, subset = NULL)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object created with the \link[=phyloseq]{phyloseq} package (must contain \code{\link[phyloseq:sample_data]{sample_data()}}).}

\item{treatment}{Column name or number, or vector of, in the \code{\link[phyloseq:sample_data]{sample_data}}.}

\item{subset}{Keyword for a subset of the treatments, can be a substring within the treatment names (e.g. 'control').}
}
\description{
Takes a \code{\link[phyloseq]{phyloseq-class}} object and finds which taxa are shared between all of the specified treatments.
}
\examples{
data(mock_phyloseq)
find_common_taxa(mock_phyloseq, treatment = 2)
find_common_taxa(mock_phyloseq, treatment = c("treatment", "day"), subset = "control")
}
\seealso{
\code{\link{find_unique_taxa}}
}
\keyword{manip}
