% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_phyloseq_generalists.R
\name{find_generalists}
\alias{find_generalists}
\alias{generalists}
\title{Find Generalists}
\usage{
find_generalists(phyloseq_obj, frequency = 0, treatment = NULL,
subset = NULL, below = FALSE, drop_samples = FALSE)
}
\arguments{
\item{phyloseq_obj}{A \code{phyloseq-class} object created with the \code{\link[phyloseq]} package.}

\item{frequency}{The minimum proportion of samples the taxa is seen in.}

\item{treatment}{Name(s) or column number(s) in the sample_data(). Function then checks if taxa seen in frequency in each treatment. If multiple sample_data() columns are given, they will be appended to the sample_data() as one column with '.' separating each.}

\item{subset}{If taxa not needed to be seen in all `treatment`, then can check only one particular treatment subset, this works for multiple treatment inputs.}

\item{below}{Does frequency define the minimum or maximum, should the presence fall below frequency or not.}

\item{drop_samples}{Should the function remove samples that that are empty after removing taxa filtered by frequency.}
}
\description{
This function takes a phyloseq object and finds which taxa are seen in a given proportion of samples, either in the entire dataset, by treatment, or a particular treatment of interest.
}
\examples{
data(mock_phyloseq)
find_generalists(mock_phyloseq, frequency = 0.3)
find_generalists(mock_phyloseq, frequency = 0.3, treatment = "day")
find_generalists(mock_phyloseq, frequency = 0.3, treatment = 3)
find_generalists(mock_phyloseq, frequency = 0.3, treatment = c("day", "treatment"))
find_generalists(mock_phyloseq, frequency = 0.3, treatment = c("day", "treatment"), subset = "soil")
find_generalists(mock_phyloseq, frequency = 0.3, treatment = c("day", "treatment"), subset = c("5","soil"))
find_generalists(mock_phyloseq, frequency = 0.3, treatment = c("day", "treatment"), subset = "5.soil")
}
\keyword{manip}
