% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_unique_genes.R
\name{find_unique_taxa}
\alias{find_unique_taxa}
\title{Find Unique Taxa}
\usage{
find_unique_taxa(phyloseq_obj, column, keyword = NULL)
}
\arguments{
\item{phyloseq_obj}{A phyloseq-class object created with the phyloseq package (must contain sample_data()).}

\item{column}{Name or column number in the sample_data(). Function then checks if taxa seen in frequency in each treatment.}

\item{keyword}{Looks for a subset of the column, can be a substring within the treatments names (e.g. 'control').}
}
\description{
This function takes a phyloseq object and finds which taxa are taxa that are unique to a specific subset of the data.
}
\examples{
data(mock_phyloseq)
find_unique_taxa(mock_phyloseq, column = 2)
}
\keyword{phyloseq}
\keyword{phylosmith}
\keyword{taxa}
\keyword{unique}
