% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_unique_taxa.R
\name{find_unique_taxa}
\alias{find_unique_taxa}
\alias{find_unique_genes}
\title{Find Unique Taxa}
\usage{
find_unique_taxa(phyloseq_obj, treatment, subset = NULL)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object created with the \link[=phyloseq]{phyloseq} package (must contain \code{\link[phyloseq:sample_data]{sample_data()}}).}

\item{treatment}{Column name or number, or vector of, in the \code{\link[phyloseq:sample_data]{sample_data}}.}

\item{subset}{Looks for a subset of the treatment, can be a substring within the treatments names (e.g. 'control').}
}
\description{
This function takes a \code{\link[phyloseq]{phyloseq-class}} object and finds which taxa are taxa that are unique to a specific subset of the data.
}
\examples{
data(mock_phyloseq)
find_unique_taxa(mock_phyloseq, treatment = 2)
find_unique_taxa(mock_phyloseq, treatment = "day")
}
\keyword{manip}
