% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_filter.R
\name{melt_phyloseq}
\alias{melt_phyloseq}
\title{Melt a phyloseq object data.table.}
\usage{
melt_phyloseq(phyloseq_obj)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object. It must contain \code{\link[phyloseq:sample_data]{sample_data()}} with information about each sample, and it must contain \code{\link[phyloseq:tax_table]{tax_table()}}) with information about each taxa/gene.}
}
\description{
melt_phyloseq takes a phyloseq object and melts its ou_table, taxa_tale, and sample_Data into a single into a data.table.
}
\seealso{
\code{\link[phyloseq:psmelt]{psmelt()}}
}
\keyword{manip}
