% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_ASVs.R
\name{merge_asvs}
\alias{merge_asvs}
\title{Merge amplicon sequence variants (asv). phylosmith}
\usage{
merge_asvs(...)
}
\arguments{
\item{...}{Any number of \code{\link[phyloseq]{phyloseq-class}} objects created with the \link[=phyloseq]{phyloseq} package (must have Amplicon Sequence Variants for calling taxa).}
}
\description{
Takes any number of phyloseq objects that have \code{\link[phyloseq:taxa_names]{taxa_names}} declared as true biological sequences or Amplicon Sequence Variants (ASVs). If a dataset has longer reads than the others, it will look for the shorter sequence within the longer and redeclare the ASV.
}
\examples{
data(mock_phyloseq); data(mock_phyloseq_2)
merge_asvs(mock_phyloseq, mock_phyloseq_2)
}
\keyword{manip}
