% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_filter.R
\name{merge_samples}
\alias{merge_samples}
\title{Merge samples with treatment groups. Function from the phylosmith-package.}
\usage{
merge_samples(phyloseq_obj, treatment, subset = NULL, merge_on = treatment)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object created with the \link[=phyloseq]{phyloseq} package.}

\item{treatment}{Column name or number, or vector of, in the \code{\link[phyloseq:sample_data]{sample_data}}, the function will look in to merge on. If you want to merge basedon treatment, assign this argument and do not assing \code{merge_on}.}

\item{subset}{If taxa not needed to be seen in all \code{treatment}, then can check only one particular treatment subset, this works for multiple treatment inputs.}

\item{merge_on}{Column name or number, or vector of, in the \code{\link[phyloseq:sample_data]{sample_data}} to merge on.}
}
\description{
This function takes a phyloseq object and merges the samples that meet the specified criteria into a single sample. This is meant for replicates, or samples statistically proven to not be significantly different and should be used with caution as it may be a misleading representation of the data.
}
\keyword{manip}
