% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mock_phyloseq_2}
\alias{mock_phyloseq_2}
\title{Mock Phyloseq Object 2}
\format{
A \code{\link[phyloseq]{phyloseq-class}} object.
}
\source{
Simulation
}
\usage{
data(mock_phyloseq_2)
}
\description{
These datasets were made-up for the purpose
of demonstrating the examples of the functions.
They are not biologically accurate, and don't
represent anything meaningful beyond being
functional phyloseq objects.
}
\keyword{datasets}
