% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{network_phyloseq}
\alias{network_phyloseq}
\title{Create an object of the abundance barplots from a phyloseq object. Function from the phylosmith-package.}
\usage{
network_phyloseq(phyloseq_obj, treatment, subset = NULL, co_occurrence_table = NULL,
classification = 'none', node_colors = 'default', cluster = FALSE,
cluster_colors = 'default', buffer = 0.5)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object created with the \link[=phyloseq]{phyloseq} package (must contain \code{\link[phyloseq:sample_data]{sample_data()}}).}

\item{treatment}{Column name or number, or vector of, in the \code{\link[phyloseq:sample_data]{sample_data}}.}

\item{subset}{If taxa not needed to be seen in all \code{treatment}, then will subset to treatments containing this string.}

\item{co_occurrence_table}{Co_Occurence table of the \code{phyloseq_obj}, computed using \code{\link{co_occurrence}}. If none is given, it will computer the co-occurrence with the \code{phyloseq_obj}, using the given \code{treatment} and $p$ = 0.05.}

\item{classification}{Column name or number in the \code{\link[phyloseq:tax_table]{tax_table}} for node colors.}

\item{node_colors}{Name of a color set from the \link[=RColorBrewer]{RColorBrewer} package or a vector palete of R accepted colors.}

\item{cluster}{if TRUE, will use igraph's \code{\link[igraph:cluster_fast_greedy]{cluster_fast_greedy}} method. Alternatively, can pass a vctor of cluster assignments with order corresponding to the order of the \code{taxa_names} in the \code{phyloseq_obj}.}

\item{cluster_colors}{Name of a color set from the \link[=RColorBrewer]{RColorBrewer} package or a vector palete of R accepted colors to use for the cluster.}

\item{buffer}{Amount of space beyond the points to extend the cluster.}
}
\description{
This function takes a \code{\link[phyloseq]{phyloseq-class}} object and creates barplots of taxa by treatment.
}
