% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{nmds_phyloseq_ggplot}
\alias{nmds_phyloseq_ggplot}
\title{Create a ggplot object of the NMDS from a phyloseq object.
Function from the phylosmith-package.}
\usage{
nmds_phyloseq_ggplot(phyloseq_obj, treatment, circle = TRUE,
labels = NULL, colors = 'default', verbose = TRUE)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object. It
must contain \code{\link[phyloseq:sample_data]{sample_data()}}) with
information about each sample, and it must contain
\code{\link[phyloseq:tax_table]{tax_table()}}) with information about each
taxa/gene.}

\item{treatment}{Column name as a string or number in the
\code{\link[phyloseq:sample_data]{sample_data}}. This can be a vector of
multiple columns and they will be combined into a new column.}

\item{circle}{Add a \code{\link[ggplot2:stat_ellipse]{stat_ellipse}} around
each of the \code{treatment} factors (\code{TRUE}).}

\item{labels}{Column name as a string or number in the
\code{\link[phyloseq:sample_data]{sample_data}} to use to place labels of
that factor instead of circle points.}

\item{colors}{Name of a color set from the
\link[=RColorBrewer]{RColorBrewer} package or a vector palete of R-accepted
colors.}

\item{verbose}{Whether or not to print the
\code{\link[vegan:metaMDS]{metaMDS}} stress convergence to console
(\code{TRUE}) or not (\code{FALSE}).}
}
\value{
ggplot-object
}
\description{
Inputs a \code{\link[phyloseq]{phyloseq-class}} object and
plots the NMDS of a treatment or set of treatments in space.
}
\examples{
nmds_phyloseq_ggplot(soil_column, c('Matrix', 'Treatment'),
circle = TRUE, verbose = FALSE)
}
