% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_occurrence.R
\name{permute_rho}
\alias{permute_rho}
\title{Permutes the pair-wise Spearman rank co-occurrence, to determine a
significant rho-cutoff. Function from the phylosmith-package.}
\usage{
permute_rho(phyloseq_obj, treatment = NULL, subset = NULL,
replicate_samples = NULL, permutations = 10, method = 'spearman',
cores = 1)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object.}

\item{treatment}{Column name as a \code{string} or \code{numeric} in the
\code{\link[phyloseq:sample_data]{sample_data}}. This can be a vector of
multiple columns and they will be combined into a new column.}

\item{subset}{A factor within the \code{treatment}. This will remove any
samples that to not contain this factor. This can be a vector of multiple
factors to subset on.}

\item{replicate_samples}{Column name as a \code{string} or \code{numeric}
in the \code{\link[phyloseq:sample_data]{sample_data}} that indicates which
samples are non-independent of each other.}

\item{permutations}{\code{numeric} Number of iterations to compute.}

\item{method}{Which correlation method to calculate, "pearson", "spearman".}

\item{cores}{\code{numeric} Number of CPU cores to use for the pair-wise
permutations. Default (0) uses max cores available. Parallelization not
available for systems running MacOS without openMP configuration.}
}
\value{
table
}
\description{
Permutes the pair-wise Spearman rank co-occurrence, to determine a
significant rho-cutoff.
}
\examples{
permute_rho(soil_column, treatment = c('Matrix', 'Treatment'),
subset = 'Amended', replicate_samples = 'Day', permutations = 1,  cores = 0)
}
\seealso{
\code{\link{co_occurrence}}
}
\keyword{nonparametric}
