% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylosmith_package.R
\docType{_PACKAGE}
\name{phylosmith}
\alias{phylosmith}
\title{phylosmith-package}
\description{
A supplementary package to build on the phyloseq package. Phyloseq objects 
are a great data-standard for microbiome and gene-expression data, this 
package is aimed to provied easy data-wrangling and visualization.
}
\details{
\url{https://schuyler-smith.github.io/phylosmith/}
}
\author{
Schuyler D. Smith \email{schuylers.emails@gmail.com}
}
\keyword{package}
