% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylosmith_package.R
\docType{package}
\name{phylosmith}
\alias{phylosmith}
\alias{phylosmith-package}
\title{phylosmith package}
\description{
phylosmith package
}
\author{
Schuyler Smith \email{sdsmith@iastate.edu}
}
\keyword{package}
