% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylosmith_package.R
\docType{package}
\name{phylosmith}
\alias{phylosmith}
\title{phylosmith-package}
\description{
phylosmith is a conglomeration of functions written to process
and analyze \code{\link[phyloseq]{phyloseq-class}} objects.
}
\details{
\url{https://schuyler-smith.github.io/phylosmith/}
}
\author{
Schuyler Smith \email{schuyler.smith@iastate.edu}
}
\keyword{package}
