% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_functions.R
\name{relative_abundance}
\alias{relative_abundance}
\title{Transform abundance data in an \code{otu_table} to relative abundance, sample-by-sample. Function from the phylosmith-package.}
\usage{
relative_abundance(phyloseq_obj)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object.}
}
\description{
Transform abundance data into relative abundance, i.e. proportional data. This is an alternative method of normalization and may not be appropriate for all datasets, particularly if your sequencing depth varies between samples.
}
\keyword{manip}
