% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_abundance.R
\name{relative_abundance}
\alias{relative_abundance}
\title{Transform abundance data in an \code{otu_table} to relative abundance, sample-by-sample.}
\usage{
relative_abundance(phyloseq_obj)
}
\arguments{
\item{phyloseq_obj}{A \code{phyloseq-class} object created with the \link[=phyloseq]{name} package.}
}
\description{
This is used to transform abundance data into relative abundance, i.e. proportional data. This is an alternative normalization.
}
\examples{
data(mock_phyloseq)
relative_abundance(mock_phyloseq)
}
\keyword{manip}
