% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{soil_column}
\alias{soil_column}
\title{Soil Column 16S Data - OTUs}
\format{
A \code{\link[phyloseq]{phyloseq-class}} object.
}
\source{
Experiment
}
\usage{
data(soil_column)
}
\description{
This dataset was from an experiment that looked at how microbes from
manure used as fertilizer move through soil systems. The manure was
applied to the soil in a controlled system, and rain simulations were
conducted. The soil was tested at various time points, as was the water
that ran through the soil (effluent).
}
\references{
\url{https://www.frontiersin.org/articles/10.3389/fmicb.2018.03197/full}
}
\keyword{datasets}
