% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_phyloseq_ggplot.R
\name{tsne_phyloseq_ggplot}
\alias{tsne_phyloseq_ggplot}
\title{Create a ggplot object using t-SNE from a phyloseq object. Function from the phylosmith-package.}
\usage{
tsne_phyloseq_ggplot(phyloseq_obj, treatment, perplexity = 10,
circle = TRUE, colors = cbcolors)
}
\arguments{
\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object created with the \code{\link[=phyloseq]{phyloseq}} package (must contain \code{\link[phyloseq:sample_data]{sample_data()}}).}

\item{treatment}{Column name or number, or vector of, in the \code{\link[phyloseq:sample_data]{sample_data}}.}

\item{perplexity}{similar to selecting the number of neighbors to consider in decision making (should not be bigger than 3 * perplexity < nrow(X) - 1, see \code{\link[=Rtsne]{Rtsne}} for interpretation)}

\item{circle}{If TRUE, add elipses around each treatment.}

\item{colors}{Name of a color set from the \link[=RColorBrewer]{RColorBrewer} package or a vector palete of R accepted colors.}
}
\description{
This function takes a \code{\link[phyloseq]{phyloseq-class}} object and plots the t-SNE of a treatment or set of treatments.
}
\seealso{
\code{\link[=Rtsne]{Rtsne}}
}
